Expose achievement revisions in UI
